[GPU CI 1/N] Init GPU CI on Anyscale (#102)
# What does this PR do?

FIrst PR in enabling GPU CI. 

This PR adds a `SkyRL-GPU` workflow that will run GPU CI on Anyscale. We
will basic unit and integration tests running on a 4xL4 instance in this
workflow. Currently, the CI will run post-merge on `main` . For pull
requests, it requires manual trigger by a maintainer. (so it should not
run on every commit even for maintainers, since that can be wasteful).


In this PR, I've only ported over one of our existing gpu tests -
`test_models.py` to run on CI . The plan is to gradually add more and
more tests after we've made them more efficient and compatible to be run
on a 4xL4 node.

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>