[Docs] Add docs for running on an existing ray cluster (#105)
# What does this PR do?

Adds documentation for running on an existing ray cluster with a
different ray version. Currently, we use a uv lockfile to pin all
dependencies, so this PR adds some instructions for overriding the ray
version. To make sure as many users as possible to through the same
"golden path", we will have one recommended Ray version that is pinned
and well-tested. In the future as we get more CI testing bandwidth, we
will expand the support to more Ray versions.

---------

Signed-off-by: SumanthRH <sumanthrh99@gmail.com>