[Fix][CI] Address generator CI test fails when model stop reason is length (#269)
Our unit test checks whether, for turns=3, the final conversation
generates 3 EOS token. This will not be the case when the model's
generation stop due to length. To address, we do the following:
- In the dummy environment, change observation `"turn {i}"` to `"give me
another solution {i}"`, which might make more sense for the model
- Increase max generate length from 1000 to 3000
- Final guard is to check when the stop reason is not `"stop"`, we don't
check the number of EOS token IDs