[fix] Fix the pretty trainer logging (#270)
Re-implement the logging fix of #250 that was reverted in #261

The issue was that using the `worker_process_setup_hook` to set logging
behavior interfered with vLLM using Ray as it's tensor parallel backend
and threw an error. vLLM apparently needs this to be unset.

Moved the logging configuration into RayPPOTrainer `init`.