[fix] Fix broken default arguments in generation entrypoint (#350)
By default we sleep() the inference engines after construction so that
we can initialize the training workers. When we created the
generation-only entrypoint, we still followed this sleep() code path,
but had to wake_up() the inference engines for generation, and wake_up()
can corrupt weights.

This PR disables `colocate_all` in the example script as a quick fix. I
will send out a (hopefully more elegant) solution to this, and add a
test to cover it