[fix] Pop and normalize OpenAI kwargs in engine init (#358)
PR #322 supported plumbing inference engine init kwargs directly to the
inference engines but, for vLLM, some of these args should not be passed
to the engine and rather be plumbed to the `OpenAIServingChat` object.

This PR pops and normalizes these OpenAI kwargs, starting with
`enable_auto_tools` and `tool_parser` (as well as their alias naming).