Support vLLM data parallelism (#360)
This PR adds support for vLLM data parallelism.

Note that we do not use the `ray` backend for the data parallel executor
in vLLM because it prevents us from assigning each vLLM worker
`num_gpus<1`, which we use to colocate inference workers with training.
Instead, we use the `mp` backend, launch `dp_size` vLLM engines, and set
vLLM's args for `data_parallel_address`, `data_parallel_rpc_port`, and
`data_parallel_rank` so that vLLM performs a handshake between the DP
groups.

Most of the other changes in this PR are just plumbing `dp_size` through
the stack (e.g., to compute `world_size`).

## Testing
* Extended the existing inference engine test to cover dp==2
* Manually ran the following configurations on 4xA100s:
  * num_inference_engines=1, tp=1, dp=4
  * num_inference_engines=1, tp=2, dp=2
  * num_inference_engines=2, tp=1, dp=2
  * num_inference_engines=4, tp=1, dp=1